[MIG] sale_order_action_invoice_create_hook: Migration to 12.0
