+++ sale_order_action_invoice_create_hook/__init__.py
@@ -1,5 +1,3 @@
-# Copyright 2017 Eficent Business and IT Consulting Services S.L.
 # License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl.html).
 
 from . import model
-from .hooks import post_load_hook
