+++ sale_order_action_invoice_create_hook/readme/CONTRIBUTORS.rst
@@ -1,0 +1,3 @@
+* Jordi Ballester Alomar <jordi.ballester@eficent.com>
+* Roser Garcia <roser.garcia@eficent.com>
+* Raf Ven <raf.ven@dynapps.be>
